requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
discord.py==2.3.2
distro==1.9.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
multidict==6.0.5
openai==1.30.2
pydantic==2.7.1
pydantic_core==2.18.2
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.11.0
yarl==1.9.4
